♻️ refactor : 색상변경, plus아이콘 크기,색상 수정,import 변경
