+++ src/assets/index.tsx
@@ -1,2 +1,17 @@
-import CloseIcon from '@/assets/icons/close.svg?react';
-export { CloseIcon };
+import CloseIcon from './icons/close.svg?react';
+import ChevronIcon from './icons/chevron.svg?react';
+import HeartFilledIcon from './icons/heart_f.svg?react';
+import HeartOutlineIcon from './icons/heart_o.svg?react';
+import BookmarkFilledIcon from './icons/bookmark_f.svg?react';
+import BookmarkOutlineIcon from './icons/bookmark_o.svg?react';
+import PlusIcon from './icons/plus.svg?react';
+
+export {
+  CloseIcon,
+  ChevronIcon,
+  HeartFilledIcon,
+  HeartOutlineIcon,
+  BookmarkFilledIcon,
+  BookmarkOutlineIcon,
+  PlusIcon,
+};
