+++ src/components/examples/HeaderExamples.tsx
@@ -4,7 +4,7 @@ import HomeHeader from "@/components/common/Header/HomeHeader";
 export default function HeaderTest() {
   return (
     <div className="min-h-screen bg-gray-950 p-4 flex flex-col gap-12">
-      <h1 className="text-title-2 text-white mb-4">🧪 Header & HomeHeader 테스트</h1>
+      <h1 className="text-title-2 text-white mb-4">Header & HomeHeader 테스트</h1>
 
       {/* HomeHeader Section */}
       <div className="border border-gray-800 rounded-md p-4">
@@ -16,7 +16,7 @@ export default function HeaderTest() {
       <div className="border border-gray-800 rounded-md p-4">
         <p className="text-caption-1 text-gray-400 mb-2">Header (뒤로가기 + 타이틀 + 하트/북마크)</p>
         <Header
-          title="테스트 타이틀"
+          title="타이틀"
           onBackClick={() => alert("🔙 뒤로가기 클릭")}
         />
       </div>
