+++ src/routes/route.tsx
@@ -19,24 +19,4 @@ const router = createBrowserRouter([
 ]);
 
 export default router;
-// import { createBrowserRouter } from 'react-router-dom';
-// import HeaderExamples from '@/components/examples/HeaderExamples.tsx';
 
-// import Example from '@/components/examples/InputExamples';
-// import TextareaExamples from '@/components/examples/TextareaExamples';
-// const router = createBrowserRouter([
-//   {
-//     path: '/Header-test',
-//     element: <HeaderExamples />,
-//   },
-//   {
-//     path: '/Input-test',
-//     element: <Example />,
-//   },
-//   {
-//     path: '/Textarea-test',
-//     element: <TextareaExamples />,
-//   }, 
-// ]);
-
-// export default router;
